+++ runtime/src/main/java/apple/avfoundation/AVPlayerItemAccessLog.java
@@ -37,7 +37,7 @@ public class AVPlayerItemAccessLog
     
     
     @Mapping("extendedLogDataStringEncoding")
-    public native @Representing("NSStringEncoding") int getExtendedLogDataStringEncoding();
+    public native @Representing("NSStringEncoding") long getExtendedLogDataStringEncoding();
     @Mapping("events")
     public native NSArray<AVPlayerItemAccessLogEvent> getEvents();
