+++ runtime/src/main/java/apple/messageui/MFMailComposeViewControllerDelegate.java
@@ -31,7 +31,7 @@ public interface MFMailComposeViewControllerDelegate
      * @since Available in iOS 3.0 and later.
      */
     @Mapping("mailComposeController:didFinishWithResult:error:")
-    void didFinish(MFMailComposeViewController controller, @Representing("MFMailComposeResult") int result, NSError error);
+    void didFinish(MFMailComposeViewController controller, @Representing("MFMailComposeResult") long result, NSError error);
     
     /*<adapter>*/
     /*</adapter>*/
